function/requirements.txt
boto3
pillow
requests
slack_bolt
